backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-dotenv==1.0.1
openai==1.45.0
psycopg2-binary==2.9.9
Pillow==10.4.0
django-environ==0.11.2
